feat: 替换 CanvasLayer2 为 CanvasLayer
