+++ examples/demos/canvas/index.ts
@@ -1,1 +1,2 @@
+export { MapRender as layer } from './layer';
 export { MapRender as event } from './event';
