+++ packages/layers/src/canvas/models/index.ts
@@ -1,8 +1,2 @@
-import CanvasModel from './canvas';
-export type CanvasModelType = 'canvas';
-
-const CanvasModels: { [key in CanvasModelType]: any } = {
-  canvas: CanvasModel,
-};
-
-export default CanvasModels;
+export * from './canvas';
+export * from './constants';
