--- packages/layers/src/canvas2/models/index.ts
@@ -1,2 +1,0 @@
-export * from './canvas';
-export * from './constants';
