+++ packages/site/docs/api/other/canvas.en.md
